requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
Werkzeug==3.0.2

# AI stack
transformers==4.44.2
torch==2.2.2
sentencepiece==0.2.0
nltk==3.9.1
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4

# Testing & tooling
pytest==8.3.2
pytest-cov==5.0.0
black==24.4.2
flake8==7.1.0
isort==5.13.2
